Add first failing tests for to_remove
Which should take a list of packages from `get_installed()` and remove both
the version returned by `get_current()` and the latest (if they're not the
same).

This is based on an existing utility that doesn't have any tests. I'm
planning to open source this as a proper gem and by having tests we can have
more confidence that it's doing the right thing and must trust to run it
from cron instead of manually.

- gds/puppet:modules/govuk/files/bin/govuk_unused_kernels

The only difference at this stage is that it doesn't support the different
suffix used by Ubuntu Lucid because we're not supporting that anymore.