Keep two latest kernels rather than one
In response to a Ubuntu kernel bug this week. If you're rebooting for
unattended-upgrades, running this utility regularly, and the latest kernel
turns out to be problematic, then you may not want the previously
known-working version to be removed.

https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1450442

The multiple calls to `reject` feels kind of ugly/slow but I think it's the
easiest to read.